requirements.txt
kafka-python
pyspark==3.5.1
pymongo
fastapi
uvicorn
requests
pandas
numpy
redis
joblib
scikit-learn
flask
pytz